+++ .devcontainer/install-typst.sh
@@ -5,3 +5,4 @@ curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
 . "$HOME/.cargo/env"
 
 cargo install --locked typst-cli
+cargo install --locked typstyle